refactor: move categoryService to services
