+++ packages/shopware-6-client/src/index.ts
@@ -2,7 +2,7 @@ import { ClientSettings, setupConfig } from "./settings";
 import { reloadConfiguration } from "./apiService";
 
 export { config } from "./settings";
-export { getCategories } from "./categoryService";
+export { getCategories } from "./services/categoryService";
 export { ProductService } from "./services/productService";
 
 export function setup(config: ClientSettings = {}): void {
